% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{create_data}
\alias{create_data}
\title{Create a grid structure of values that contain the given ratios}
\usage{
create_data(ratios = NULL, ...)
}
\arguments{
\item{ratios}{A numeric vector of ratios to use}
}
\value{
A data frame
}
\description{
Create a grid structure of values that contain the given ratios
}
