% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{create_grid}
\alias{create_grid}
\title{Create a grid with ratio inputs at random locations}
\usage{
create_grid(inputs, dims = c(10, 10))
}
\arguments{
\item{inputs}{Ratios to be placed on a grid}

\item{dims}{Number of units for x and y grid coordinates}

\item{...}{Additional args}
}
\value{
A data frame
}
\description{
Create a grid with ratio inputs at random locations
}
