% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_lineup.R
\name{create_lineup}
\alias{create_lineup}
\title{Create a study lineup}
\usage{
create_lineup(
  ratio = NULL,
  plot = NULL,
  wp.size = 5,
  output = c("both", "design", "data")
)
}
\arguments{
\item{ratio}{A numeric vector of ratios to test}

\item{plot}{A character vector of plot types}

\item{wp.size}{Whole-plot block size}

\item{output}{Tell the function to produce the design, data, or both}
}
\value{
A dataframe
}
\description{
Create a study lineup
}
