% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{design}
\alias{design}
\title{STAT 218 Heatmap Design}
\format{
## `design`
A data frame with 360 rows and 3 columns:
/describe{
\item{block}{Treatment conditions for a single particiapnt}
\item{trt.ratio}{Ratio of bars}
\item{plot}{Plot type (2D, 3D digital with color, 3D digital with solid color, 3D print)}
...
}
}
\usage{
design
}
\description{
Design used for administering the experiment to STAT 218 at
University of Nebraska-Lincoln ...
}
\keyword{datasets}
