% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{design_data}
\alias{design_data}
\title{STAT 218 Heatmap Data}
\format{
## `design_data`
A data frame with 3,600 rows and 8 columns:
/describe{
\item{plot}{Plot type (2D, 3D digital with color, 3D digital with solid color, 3D print)}
\item{plotID}{Numeric identifier of plot}
\item{x, y}{Grid coordinate positions}
\item{label}{Label to identify coordinate of interest}
\item{ratio}{Ratio comparison}
\item{bar}{Smaller or larger bar of ratio comparison}
\item{z}{Value of tested unit}
...
}
}
\usage{
design_data
}
\description{
Data used to generate plots for the experiment given to STAT 218 students
at University of Nebraska-Lincoln ...
}
\keyword{datasets}
