% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{grid_distribution}
\alias{grid_distribution}
\title{Generate a density around a coordinate location}
\usage{
grid_distribution(x, y, z, span = 2, mult = 1, ...)
}
\arguments{
\item{x}{Coordinate on x-axis}

\item{y}{Coordinate on y-axis}

\item{z}{Height of target bar}

\item{span}{Number of grid spaces to span over in the x and y direction}

\item{mult}{Additional multiplier to surrounding bar. By default, a value of 1 uses the inverse of the Euclidean distance on grid coordinates with respect to the given x and y inputs.}

\item{...}{Additional args to pass to `rnorm`}
}
\value{
A data frame
}
\description{
Generate a density around a coordinate location
}
